@nseinlet 2023-11-15
@rco-odoo is the original author, I was just spreading the news.