@nseinlet 2023-11-15
You could probably comment with an explain plan before/after. My 2 cents.